requirements_full.txt
# Generated requirements for CryoProtect
# External packages detected from imports
ChEMBL_CryoProtectants_Supabase
ChEMBL_Integrated_Import
Crypto
IPython
OpenSSL
PubChem_CryoProtectants_Supabase_Enhanced
add_performance_indexes
aiohttp
alembic
aniso8601
api
apispec
app
auth_config
backup
batch_utils
batch_utils_mock
blinker
boto3
botocore
bottle
certifi
charset_normalizer
chembl
chembl_import
chembl_search_utils
chembl_webresource_client
chemical_data
click
client
cloud_storage
codex_cli_wrapper
colorama
config
connection_pool_wrapper
coverage
cryoprotectant_identifiers
cryptography
database
db_connection_utils
db_connection_utils_mock
distutils
django
docutils
python-dotenv
ecs_logging
enhance_pubchem_properties
execute_rls_sql_via_mcp
falcon
file_backup
fix_database_modular
Flask
flask_apispec
flask_cors
flask_limiter
flask_mail
flask_restful
h2
idna
import_full_chembl
import_pubchem_data_chunked
import_reference_compounds
ip_resolver
itsdangerous
Jinja2
js
PyJWT
logging_config
logging_enhanced
markupsafe
marshmallow
matplotlib
mock_mcp_tool
monitoring
monitoring_utils
notify
ntlm
numpy
optimize_toxicity_queries
optimized_connection_pool
packaging
pandas
pip
populate_database
postgres_direct
postgres_helper
postgrest
prometheus_client
property_utils
property_utils_mock
psutil
psycopg2-binary
pubchem
pyodide
pyramid
pytest
pythonjsonlogger
pytz
rate_limit_config
rdkit
reconcile_chembl_properties
redis
reportlab
requests
rls_utils
schedule
scipy
seaborn
security
security_headers
service_role_helper
six
scikit-learn
sphinx
sql_executor
SQLAlchemy
supabase
supabase_adapter
supabase_database_audit
supabase_direct
supabase_direct_connection
supabase_mcp_tools
supabase_postgres_helper
tabulate
test_cache
test_client
test_rate_limiter
test_scheduler
test_utils
tests
tornado
transaction_utils
transaction_utils_mock
urllib3
use_mcp_tool
verify_chembl_data
verify_chembl_integrity
verify_imported_data
webargs
webob
webtest
Werkzeug
xlsxwriter
PyYAML

# Additional known required packages
flask-apispec
flask-cors
flask-mail
flask-restful
prometheus-client
